# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MenuItemCategory = typeof MenuItemCategory[keyof typeof MenuItemCategory];


export const MenuItemCategory = {
  Makanan: 'Makanan',
  Minuman: 'Minuman',
} as const;

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: MenuItemCategory;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
  isDefault: boolean;
  orderIndex: number;
  createdAt: string;
}

export type MenuInputCategory = typeof MenuInputCategory[keyof typeof MenuInputCategory];


export const MenuInputCategory = {
  Makanan: 'Makanan',
  Minuman: 'Minuman',
} as const;

export interface MenuInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  category: MenuInputCategory;
  imageUrl?: string;
  isAvailable?: boolean;
}

export type MenuUpdateCategory = typeof MenuUpdateCategory[keyof typeof MenuUpdateCategory];


export const MenuUpdateCategory = {
  Makanan: 'Makanan',
  Minuman: 'Minuman',
} as const;

export interface MenuUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  category?: MenuUpdateCategory;
  imageUrl?: string;
  isAvailable?: boolean;
}

export type MenuSummaryByCategory = {
  Makanan: number;
  Minuman: number;
};

export interface MenuSummary {
  total: number;
  available: number;
  byCategory: MenuSummaryByCategory;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: string;
  totalPrice: number;
  finalPrice: number;
  /** @nullable */
  voucherCode?: string | null;
  discountRate: number;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  customerAddress: string;
  items: string;
  totalPrice: number;
  finalPrice: number;
  voucherCode?: string;
  discountRate?: number;
  paymentMethod: string;
  notes?: string;
}

export interface RevenueSummary {
  totalRevenue: number;
  totalOrders: number;
  totalDiscount: number;
  ordersWithVoucher: number;
}

export interface DeleteResult {
  success: boolean;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  message: string;
}

export type ListMenusParams = {
category?: string;
available?: boolean;
};