# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cafe App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menus
    description: Menu management
  - name: orders
    description: Order management
  - name: admin
    description: Admin authentication
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menus:
    get:
      operationId: listMenus
      tags: [menus]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenu
      tags: [menus]
      summary: Create a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: Bad request

  /menus/summary:
    get:
      operationId: getMenuSummary
      tags: [menus]
      summary: Get menu summary stats (count per category, total, available count)
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuSummary"

  /menus/{id}:
    get:
      operationId: getMenu
      tags: [menus]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    patch:
      operationId: updateMenu
      tags: [menus]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    delete:
      operationId: deleteMenu
      tags: [menus]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request

  /orders/revenue:
    get:
      operationId: getOrderRevenue
      tags: [orders]
      summary: Get revenue and order summary stats
      responses:
        "200":
          description: Revenue summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueSummary"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized


components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: integer
        category:
          type: string
          enum: [Makanan, Minuman]
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isDefault:
          type: boolean
        orderIndex:
          type: integer
        createdAt:
          type: string
      required: [id, name, price, category, isAvailable, isDefault, orderIndex, createdAt]

    MenuInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: integer
          minimum: 0
        category:
          type: string
          enum: [Makanan, Minuman]
        imageUrl:
          type: string
        isAvailable:
          type: boolean
      required: [name, price, category]

    MenuUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: integer
          minimum: 0
        category:
          type: string
          enum: [Makanan, Minuman]
        imageUrl:
          type: string
        isAvailable:
          type: boolean

    MenuSummary:
      type: object
      properties:
        total:
          type: integer
        available:
          type: integer
        byCategory:
          type: object
          properties:
            Makanan:
              type: integer
            Minuman:
              type: integer
          required: [Makanan, Minuman]
      required: [total, available, byCategory]

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        items:
          type: string
        totalPrice:
          type: integer
        finalPrice:
          type: integer
        voucherCode:
          type: ["string", "null"]
        discountRate:
          type: integer
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, customerName, customerPhone, customerAddress, items, totalPrice, finalPrice, discountRate, paymentMethod, createdAt]

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 1
        customerPhone:
          type: string
          minLength: 1
        customerAddress:
          type: string
          minLength: 1
        items:
          type: string
        totalPrice:
          type: integer
        finalPrice:
          type: integer
        voucherCode:
          type: string
        discountRate:
          type: integer
        paymentMethod:
          type: string
        notes:
          type: string
      required: [customerName, customerPhone, customerAddress, items, totalPrice, finalPrice, paymentMethod]

    RevenueSummary:
      type: object
      properties:
        totalRevenue:
          type: integer
        totalOrders:
          type: integer
        totalDiscount:
          type: integer
        ordersWithVoucher:
          type: integer
      required: [totalRevenue, totalOrders, totalDiscount, ordersWithVoucher]

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    AdminLoginInput:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required: [password]

    AdminLoginResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]