# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}