# lib/api-zod/src/generated/types/adminLoginResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginResult {
  success: boolean;
  message: string;
}