# lib/api-zod/src/generated/types/menuInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuInputCategory } from './menuInputCategory';

export interface MenuInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  category: MenuInputCategory;
  imageUrl?: string;
  isAvailable?: boolean;
}