# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemCategory } from './menuItemCategory';

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: MenuItemCategory;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
  isDefault: boolean;
  orderIndex: number;
  createdAt: string;
}