# lib/api-zod/src/generated/types/menuSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuSummaryByCategory } from './menuSummaryByCategory';

export interface MenuSummary {
  total: number;
  available: number;
  byCategory: MenuSummaryByCategory;
}