# lib/api-zod/src/generated/types/menuUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuUpdateCategory } from './menuUpdateCategory';

export interface MenuUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  category?: MenuUpdateCategory;
  imageUrl?: string;
  isAvailable?: boolean;
}