# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: string;
  totalPrice: number;
  finalPrice: number;
  /** @nullable */
  voucherCode?: string | null;
  discountRate: number;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}