# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Cafe App API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  customerAddress: string;
  items: string;
  totalPrice: number;
  finalPrice: number;
  voucherCode?: string;
  discountRate?: number;
  paymentMethod: string;
  notes?: string;
}